relatorio-d4sign/requirements.txt
Flask
requests
python-dotenv
Flask-Cors